Split api stream structural modules
